feat: RecordList 그리기 Cell 구현
